Implement sorting by fetching metadata first
Useful if we want to combine sorting and paging. Kubernetes does not
support sorting on the server side, therefore the combination of sorting
and paging should takes place on the client. However, sorting can be
only performed on the full set of data and fetching all objects in order
to sort them is expensive memory and network wise.

As an attempt to reduce the network traffic and consumed memory, we
leverage that all Korifi objects have labels that match sort criteria.
Then, we first list the `PartialObjectMetadata` for the object (it
contains labels), sort it, maybe page it and "explode" the metadata page
into a page of full korifi objects. This explosion is implemented as a
subsequent list by concrete guids, and order the result to match the
sorted metadata.

An alternative approach we looked into is to start a manager cache in
the api that only caches metadata (without managed fields).

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>