Do not fail if there is no object for a guid
When listing objects with the descriptor client, the klient would
perform the following sequence of actions:
1. Collect guids for objects that match label selectors
2. Order and page the guids
3. Map guids to objects

In a real deployment where objects are being created asynchronously, it
could happen that after guids have been fetched, objects for few of the
guids to have been deleted in the meanwhile. If the mapper returns an
error on such an occasion, this would cause the list request to fail.

This commit changes the object mapper to ignore this error and simply
skip the object that is gone.

A more correct solution would be to go back to step 1 and retry list
until all guids are succefully mapped. However, we believe that such a
complexity would be an overkill so we opted for the simpler solution

See CI failure: https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-e2es-periodic/builds/21600#L682b501b:2266

Co-authored-by: Danail Branekov <danailster@gmail.com>