Paging and Ordering for ServiceBindings
ServiceBinding's `DisplayName` are optional and annotated as
`omitempty`. This results into not populating the respective column in
the ETCD (i.e. leaving it `nil`)

In order to support sorting by such columns, we default nil columns to
the respective zero value

**NOTE** that the display name of the binding MUST be `*string` (and not
plain `string`) because:
a) we need to support manifests with bindings that have their name
omitted
b) the `binding_name` field in `VCAP_SERVICES` should be `null` for
unnamed bindings (and NOT `""`)

issue #3701

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>