Replace `gopkg.in/yaml.v3` with `go.yaml.in/yaml/v3`
The [`gopkg.in/yaml.v3`](https://github.com/go-yaml/yaml) library seems
to be no longer maintained (see its README). One of our dependencies
(`k8s.io/kube-openapi`) has already
[switched](https://github.com/kubernetes/kube-openapi/commit/e9823e99808e8e91a64f46f2541bf3434c02e3a2)
to `go.yaml.in/yaml/v3` and that breaks Korifi.

Aparently, we need to move on and switch to the alternative yaml library
- [go.yaml.in/yaml/v3`](https://github.com/yaml/go-yaml) which claims to
  be a supported fork of the initial library.

Sadly, switching to the maintained fork would not be enough. We have
other dependencies (such as gomega and zap) that need to implement the
switch as well.