Remove the upload-gh-pages workflow
It is not working as expected. In particular it does not get triggered
when commits get pushed to the gh-pages branch

There seems to be a "Deploy from a branch" option that does just that,
so we can get back to that, which makes the workflow redundant.

The workflow was introduced because we believed that the helmchart
content is not uploaded as static html cotent, but that proved to be
wrong