Fix empty errors in logs when a reconciled resource is not ready
`NotReadyError` ensures that its error message is not empty.

The initial issue was that when the app droplet is not set, the error
gets its reson set to `DropletNotAssigned`, but the reason is not
reflected in the error message. This commit includes the reason, and
ensures that the message is at least `resource not ready`

While being here, do not requeue reconcile when the app droplet is not
set in its spec. It makes no sense to requeue the reconcile event as it
is expected that the droplet is set on the spec externally. When that
happens the app would be reconciled anyway

fixes #3862